Bug 641025 - Incremental GC (r=igor,smaug,roc,cdleary,gregor)
UltraBlame original commit: 2a8ceeb27f7c8a6a43ef01d55d827ddccb887e33

+++ js/src/methodjit/Compiler.h
@@ -484,7 +484,7 @@ private:
     bool hasGlobalReallocation;
     bool oomInVector;       
     bool overflowICSpace;   
-    uint32_t gcNumber;
+    uint64_t gcNumber;
     enum { NoApplyTricks, LazyArgsObj } applyTricks;
     PCLengthEntry *pcLengths;
